# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PostEntity and Comments API
  description: |
    RESTful API for managing posts, comments, post details, and tags.
    
    This API allows you to:
    - Create, read, update, and delete posts
    - Add comments to posts
    - Manage post details and tags
    - Associate tags with posts
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8081
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /api/posts:
    get:
      tags:
        - Posts
      summary: Get all postEntities
      description: Retrieves a list of all postEntities
      operationId: getAllPosts
      responses:
        '200':
          description: Successfully retrieved postEntities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostListResponse'
              example:
                status: 200
                message: "Posts retrieved successfully"
                success: true
                payload:
                  - id: 1
                    title: "Introduction to Spring Boot"
                    createdAt: "2025-12-02T10:00:00"
                    updatedAt: "2025-12-02T10:00:00"
                    comments: []
                    tags: []
                    postDetails: null
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Posts
      summary: Create a new postEntity
      description: Creates a new postEntity with the provided information
      operationId: createPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostRequest'
            example:
              title: "Introduction to Spring Boot"
              tagIds: [1, 2]
      responses:
        '201':
          description: PostEntity created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponseWrapper'
              example:
                status: 201
                message: "PostEntity created successfully"
                success: true
                payload:
                  id: 1
                  title: "Introduction to Spring Boot"
                  createdAt: "2025-12-02T10:00:00"
                  updatedAt: "2025-12-02T10:00:00"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/posts/{id}:
    get:
      tags:
        - Posts
      summary: Get postEntity by ID
      description: Retrieves a specific postEntity by its ID
      operationId: getPostById
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '200':
          description: Successfully retrieved postEntity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponseWrapper'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Posts
      summary: Update a postEntity
      description: Updates an existing postEntity with the provided information
      operationId: updatePost
      parameters:
        - $ref: '#/components/parameters/PostId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostRequest'
            example:
              title: "Updated PostEntity Title"
              tagIds: [1, 3]
      responses:
        '200':
          description: PostEntity updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponseWrapper'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Posts
      summary: Delete a postEntity
      description: Deletes a postEntity by its ID
      operationId: deletePost
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '204':
          description: PostEntity deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Comments endpoints
  /api/comments:
    get:
      tags:
        - Comments
      summary: Get all comments
      description: Retrieves a list of all comments
      operationId: getAllComments
      responses:
        '200':
          description: Successfully retrieved comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostComment'
              example:
                - id: 1
                  review: "Great postEntity!"
                  post:
                    id: 1
                  createdAt: "2025-12-02T10:30:00"
                  updatedAt: "2025-12-02T10:30:00"

    post:
      tags:
        - Comments
      summary: Create a new comment
      description: Creates a new comment for a postEntity
      operationId: createComment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostComment'
            example:
              review: "This is a great postEntity!"
              post:
                id: 1
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostComment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/comments/{id}:
    get:
      tags:
        - Comments
      summary: Get comment by ID
      description: Retrieves a specific comment by its ID
      operationId: getCommentById
      parameters:
        - name: id
          in: path
          required: true
          description: Comment ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostComment'
        '404':
          description: Comment not found

    put:
      tags:
        - Comments
      summary: Update a comment
      description: Updates an existing comment
      operationId: updateComment
      parameters:
        - name: id
          in: path
          required: true
          description: Comment ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostComment'
      responses:
        '200':
          description: Comment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostComment'
        '404':
          description: Comment not found

    delete:
      tags:
        - Comments
      summary: Delete a comment
      description: Deletes a comment by its ID
      operationId: deleteComment
      parameters:
        - name: id
          in: path
          required: true
          description: Comment ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Comment deleted successfully
        '404':
          description: Comment not found
  /api/comments/post/{postId}:
    get:
      tags:
        - Comments
      summary: Get comments by postEntity ID
      description: Retrieves all comments for a specific postEntity
      operationId: getCommentsByPostId
      parameters:
        - name: postId
          in: path
          required: true
          description: PostEntity ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostComment'

  # PostEntity Details endpoints
...
  /api/post-details/{id}:
    get:
      tags:
        - PostEntity Details
      summary: Get postEntity details by ID
      description: Retrieves postEntity details by its ID
      operationId: getPostDetailsById
      parameters:
        - name: id
          in: path
          required: true
          description: PostEntity details ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved postEntity details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostDetails'
        '404':
          description: PostEntity details not found

    put:
      tags:
        - PostEntity Details
      summary: Update postEntity details
      description: Updates existing postEntity details
      operationId: updatePostDetails
      parameters:
        - name: id
          in: path
          required: true
          description: PostEntity details ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostDetails'
      responses:
        '200':
          description: PostEntity details updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostDetails'
        '404':
          description: PostEntity details not found

    delete:
      tags:
        - PostEntity Details
      summary: Delete postEntity details
      description: Deletes postEntity details by its ID
      operationId: deletePostDetails
      parameters:
        - name: id
          in: path
          required: true
          description: PostEntity details ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: PostEntity details deleted successfully
        '404':
          description: PostEntity details not found
  /api/post-details/post/{postId}:
    get:
      tags:
        - PostEntity Details
      summary: Get postEntity details by postEntity ID
      description: Retrieves postEntity details for a specific postEntity
      operationId: getPostDetailsByPostId
      parameters:
        - name: postId
          in: path
          required: true
          description: PostEntity ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved postEntity details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostDetails'
        '404':
          description: PostEntity details not found

  # Tags endpoints
...
components:
  parameters:
    PostId:
      name: id
      in: path
      required: true
      description: PostEntity ID
      schema:
        type: integer
        format: int64
        example: 1
  schemas:
    PostRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: The title of the postEntity
          example: "Introduction to Spring Boot"
        tagIds:
          type: array
          items:
            type: integer
            format: int64
          description: Array of tag IDs to associate with the postEntity
          example: [1, 2]
    PostResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the postEntity
          example: 1
        title:
          type: string
          description: The title of the postEntity
          example: "Introduction to Spring Boot"
        postDetails:
          $ref: '#/components/schemas/PostDetailsResponse'
          nullable: true
        comments:
          type: array
          items:
            $ref: '#/components/schemas/PostCommentResponse'
          description: List of comments for this postEntity
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagResponse'
          description: List of tagEntities associated with this postEntity
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the postEntity was created
          example: "2025-12-02T10:00:00"
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the postEntity was last updated
          example: "2025-12-02T10:00:00"
    PostListResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        message:
          type: string
          example: "Posts retrieved successfully"
        success:
          type: boolean
          example: true
        payload:
          type: array
          items:
            $ref: '#/components/schemas/PostResponse'
    PostResponseWrapper:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 201
        message:
          type: string
          example: "PostEntity created successfully"
        success:
          type: boolean
          example: true
        payload:
          $ref: '#/components/schemas/PostResponse'
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          example: "/api/postEntities"
    PostComment:
      type: object
      required:
        - review
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the comment
          example: 1
        review:
          type: string
          description: The comment/review text
          example: "Great postEntity!"
        post:
          type: object
          properties:
            id:
              type: integer
              format: int64
          description: The postEntity this comment belongs to
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the comment was created
          example: "2025-12-02T10:30:00"
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the comment was last updated
          example: "2025-12-02T10:30:00"
    PostCommentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        review:
          type: string
          example: "Great postEntity!"
        postId:
          type: integer
          format: int64
          example: 1
        createdAt:
          type: string
          format: date-time
          example: "2025-12-02T10:30:00"
        updatedAt:
          type: string
          format: date-time
          example: "2025-12-02T10:30:00"
    PostDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the postEntity details
          example: 1
        description:
          type: string
          description: Detailed description of the postEntity
          maxLength: 5000
          example: "This is a detailed description of the postEntity content"
        post:
          type: object
          properties:
            id:
              type: integer
              format: int64
          description: The postEntity these details belong to
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the details were created
          example: "2025-12-02T10:00:00"
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the details were last updated
          example: "2025-12-02T10:00:00"
    PostDetailsResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        postId:
          type: integer
          format: int64
          example: 1
        description:
          type: string
          example: "This is a detailed description"
        createdAt:
          type: string
          format: date-time
          example: "2025-12-02T10:00:00"
        updatedAt:
          type: string
          format: date-time
          example: "2025-12-02T10:00:00"
...
    TagResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Spring Boot"
    Error:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 400
        message:
          type: string
          example: "Bad Request"
        success:
          type: boolean
          example: false
        errors:
          type: array
          items:
            type: string
          example: ["Title is required"]
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          example: "/api/postEntities"
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            status: 400
            message: "Bad Request"
            success: false
            errors: ["Title is required"]
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            status: 404
            message: "PostEntity not found"
            success: false
            errors: ["PostEntity with id 1 not found"]
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            status: 500
            message: "Internal Server Error"
            success: false
            errors: ["An unexpected error occurred"]